@@ -2,6 +2,8 @@ package com.example.springdata;
 
 import java.util.Arrays;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -10,21 +12,28 @@ import org.springframework.context.annotation.Profile;
 
 @SpringBootApplication
 public class TodoApplication {
+    private final Logger logger = LoggerFactory.getLogger(TodoApplication.class);
 
     public static void main(String[] args) {
         SpringApplication.run(TodoApplication.class, args);
     }
 
-    @Profile("dev")
+    @Profile("!prod")
     @Bean
     public CommandLineRunner init(TodoRepository repository) {
         return args -> {
-            var entities = Arrays.asList(
-                    new Todo("configuration", "congratulations, you have set up JPA correctly!", false),
-                    new Todo("test", "check if this is working", false),
-                    new Todo("demo", "show to friends and family", false));
-            repository.deleteAll();
-            repository.saveAll(entities);
+            if (repository.count() == 0) {
+                logger.info("Seeding database");
+                var entities = Arrays.asList(
+                        new Todo("configuration",
+                                "congratulations, you have set up your Azure Container App correctly!",
+                                false),
+                        new Todo("test", "check if this is working", false),
+                        new Todo("demo", "show to friends and family", false));
+                repository.saveAll(entities);
+            } else {
+                logger.info("Database already seeded");
+            }
         };
     }
 }
